requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.2
djangorestframework==3.15.1
flake8==7.0.0
gunicorn==22.0.0
mccabe==0.7.0
packaging==24.0
pillow==10.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
setuptools==69.5.1
sqlparse==0.5.0
typing_extensions==4.11.0
